+++ backend/database/migrations/2026_05_20_000005_create_detalle_pedido_table.php
@@ -11,9 +11,9 @@ return new class extends Migration
         if (Schema::hasTable('detalle_pedido')) return;
         Schema::create('detalle_pedido', function (Blueprint $table) {
             $table->id('id_detalle_pedido');
-            $table->integer('id_pedido');
+            $table->unsignedBigInteger('id_pedido');
             $table->foreign('id_pedido')->references('id_pedido')->on('pedidos');
-            $table->integer('id_producto');
+            $table->unsignedBigInteger('id_producto');
             $table->foreign('id_producto')->references('id_producto')->on('productos');
             $table->integer('cantidad');
             $table->decimal('subtotal', 10, 2);
